Reorg code
